updated the player service
PlayerRepository: at the getPlayerData function edited it so that it won't take a fixed parameters.

Player Model: made the player name required and unique, and made the email unique.

PlayerService: added two functions to get a player by it's name/email, removed the createGuestPlayer and upgradeToFullAccount and integrated it's logic into the createPlayer function, fixed the updateOverTimeStats function.