implemented the authentication logix
added two middleware's to check whether the player is logged in or no.

added the AuthRoute which handles player registers as guest or full account, upgrade a guest account to full, login, logout and is the player logged in.

add the JWTHelper which includes JWT related function, verifying a token, generating a new token.

added the Validator class which helps in reducing the code required to validate the data.

in the PlayerService updated the createPlayer and added the upgradeToFullAccount function.

in the PlayerRepository fixed the setPlayerData function and updated the getPlayerData so that it now can be filterd with any paramter.